+++ google-beta/provider.go
@@ -116,6 +116,11 @@ import (
 	"github.com/hashicorp/terraform-provider-google-beta/google-beta/services/workflows"
 	"github.com/hashicorp/terraform-provider-google-beta/google-beta/services/workstations"
 
+	"github.com/hashicorp/terraform-provider-google-beta/google-beta/services/composer"
+	"github.com/hashicorp/terraform-provider-google-beta/google-beta/services/container"
+	"github.com/hashicorp/terraform-provider-google-beta/google-beta/services/containeraws"
+	"github.com/hashicorp/terraform-provider-google-beta/google-beta/services/containerazure"
+	"github.com/hashicorp/terraform-provider-google-beta/google-beta/services/servicenetworking"
 	"github.com/hashicorp/terraform-provider-google-beta/google-beta/tpgiamresource"
 	transport_tpg "github.com/hashicorp/terraform-provider-google-beta/google-beta/transport"
 	"github.com/hashicorp/terraform-provider-google-beta/google-beta/verify"
@@ -806,15 +811,15 @@ func DatasourceMapWithErrors() (map[string]*schema.Resource, error) {
 		"google_billing_account":                              billing.DataSourceGoogleBillingAccount(),
 		"google_bigquery_default_service_account":             bigquery.DataSourceGoogleBigqueryDefaultServiceAccount(),
 		"google_cloudbuild_trigger":                           cloudbuild.DataSourceGoogleCloudBuildTrigger(),
-		"google_cloudfunctions_function":                      DataSourceGoogleCloudFunctionsFunction(),
+		"google_cloudfunctions_function":                      cloudfunctions.DataSourceGoogleCloudFunctionsFunction(),
 		"google_cloudfunctions2_function":                     cloudfunctions2.DataSourceGoogleCloudFunctions2Function(),
 		"google_cloud_asset_resources_search_all":             DataSourceGoogleCloudAssetResourcesSearchAll(),
 		"google_cloud_identity_groups":                        cloudidentity.DataSourceGoogleCloudIdentityGroups(),
 		"google_cloud_identity_group_memberships":             cloudidentity.DataSourceGoogleCloudIdentityGroupMemberships(),
-		"google_cloud_run_locations":                          DataSourceGoogleCloudRunLocations(),
+		"google_cloud_run_locations":                          cloudrun.DataSourceGoogleCloudRunLocations(),
 		"google_cloud_run_service":                            cloudrun.DataSourceGoogleCloudRunService(),
-		"google_composer_environment":                         DataSourceGoogleComposerEnvironment(),
-		"google_composer_image_versions":                      DataSourceGoogleComposerImageVersions(),
+		"google_composer_environment":                         composer.DataSourceGoogleComposerEnvironment(),
+		"google_composer_image_versions":                      composer.DataSourceGoogleComposerImageVersions(),
 		"google_compute_address":                              compute.DataSourceGoogleComputeAddress(),
 		"google_compute_addresses":                            compute.DataSourceGoogleComputeAddresses(),
 		"google_compute_backend_service":                      compute.DataSourceGoogleComputeBackendService(),
@@ -852,14 +857,14 @@ func DatasourceMapWithErrors() (map[string]*schema.Resource, error) {
 		"google_compute_subnetwork":                           compute.DataSourceGoogleComputeSubnetwork(),
 		"google_compute_vpn_gateway":                          compute.DataSourceGoogleComputeVpnGateway(),
 		"google_compute_zones":                                compute.DataSourceGoogleComputeZones(),
-		"google_container_azure_versions":                     DataSourceGoogleContainerAzureVersions(),
-		"google_container_aws_versions":                       DataSourceGoogleContainerAwsVersions(),
-		"google_container_attached_versions":                  DataSourceGoogleContainerAttachedVersions(),
-		"google_container_attached_install_manifest":          DataSourceGoogleContainerAttachedInstallManifest(),
-		"google_container_cluster":                            DataSourceGoogleContainerCluster(),
-		"google_container_engine_versions":                    DataSourceGoogleContainerEngineVersions(),
-		"google_container_registry_image":                     DataSourceGoogleContainerImage(),
-		"google_container_registry_repository":                DataSourceGoogleContainerRepo(),
+		"google_container_azure_versions":                     containerazure.DataSourceGoogleContainerAzureVersions(),
+		"google_container_aws_versions":                       containeraws.DataSourceGoogleContainerAwsVersions(),
+		"google_container_attached_versions":                  containerattached.DataSourceGoogleContainerAttachedVersions(),
+		"google_container_attached_install_manifest":          containerattached.DataSourceGoogleContainerAttachedInstallManifest(),
+		"google_container_cluster":                            container.DataSourceGoogleContainerCluster(),
+		"google_container_engine_versions":                    container.DataSourceGoogleContainerEngineVersions(),
+		"google_container_registry_image":                     containeranalysis.DataSourceGoogleContainerImage(),
+		"google_container_registry_repository":                containeranalysis.DataSourceGoogleContainerRepo(),
 		"google_dataproc_metastore_service":                   dataprocmetastore.DataSourceDataprocMetastoreService(),
 		"google_datastream_static_ips":                        DataSourceGoogleDatastreamStaticIps(),
 		"google_game_services_game_server_deployment_rollout": gameservices.DataSourceGameServicesGameServerDeploymentRollout(),
@@ -918,7 +923,7 @@ func DatasourceMapWithErrors() (map[string]*schema.Resource, error) {
 		"google_sql_database":                                 sql.DataSourceSqlDatabase(),
 		"google_sql_database_instance":                        DataSourceSqlDatabaseInstance(),
 		"google_sql_database_instances":                       DataSourceSqlDatabaseInstances(),
-		"google_service_networking_peered_dns_domain":         DataSourceGoogleServiceNetworkingPeeredDNSDomain(),
+		"google_service_networking_peered_dns_domain":         servicenetworking.DataSourceGoogleServiceNetworkingPeeredDNSDomain(),
 		"google_storage_bucket":                               DataSourceGoogleStorageBucket(),
 		"google_storage_bucket_object":                        DataSourceGoogleStorageBucketObject(),
 		"google_storage_bucket_object_content":                DataSourceGoogleStorageBucketObjectContent(),
@@ -1645,8 +1650,8 @@ func ResourceMapWithErrors() (map[string]*schema.Resource, error) {
 			"google_bigtable_instance":                      bigtable.ResourceBigtableInstance(),
 			"google_bigtable_table":                         bigtable.ResourceBigtableTable(),
 			"google_billing_subaccount":                     resourcemanager.ResourceBillingSubaccount(),
-			"google_cloudfunctions_function":                ResourceCloudFunctionsFunction(),
-			"google_composer_environment":                   ResourceComposerEnvironment(),
+			"google_cloudfunctions_function":                cloudfunctions.ResourceCloudFunctionsFunction(),
+			"google_composer_environment":                   composer.ResourceComposerEnvironment(),
 			"google_compute_attached_disk":                  compute.ResourceComputeAttachedDisk(),
 			"google_compute_instance":                       compute.ResourceComputeInstance(),
 			"google_compute_disk_async_replication":         compute.ResourceComputeDiskAsyncReplication(),
@@ -1666,9 +1671,9 @@ func ResourceMapWithErrors() (map[string]*schema.Resource, error) {
 			"google_compute_shared_vpc_host_project":        compute.ResourceComputeSharedVpcHostProject(),
 			"google_compute_shared_vpc_service_project":     compute.ResourceComputeSharedVpcServiceProject(),
 			"google_compute_target_pool":                    compute.ResourceComputeTargetPool(),
-			"google_container_cluster":                      ResourceContainerCluster(),
-			"google_container_node_pool":                    ResourceContainerNodePool(),
-			"google_container_registry":                     ResourceContainerRegistry(),
+			"google_container_cluster":                      container.ResourceContainerCluster(),
+			"google_container_node_pool":                    container.ResourceContainerNodePool(),
+			"google_container_registry":                     containeranalysis.ResourceContainerRegistry(),
 			"google_dataflow_job":                           ResourceDataflowJob(),
 			"google_dataflow_flex_template_job":             ResourceDataflowFlexTemplateJob(),
 			"google_dataproc_cluster":                       ResourceDataprocCluster(),
@@ -1693,7 +1698,7 @@ func ResourceMapWithErrors() (map[string]*schema.Resource, error) {
 			"google_logging_project_bucket_config":          ResourceLoggingProjectBucketConfig(),
 			"google_monitoring_dashboard":                   ResourceMonitoringDashboard(),
 			"google_project_service_identity":               resourcemanager.ResourceProjectServiceIdentity(),
-			"google_service_networking_connection":          ResourceServiceNetworkingConnection(),
+			"google_service_networking_connection":          servicenetworking.ResourceServiceNetworkingConnection(),
 			"google_sql_database_instance":                  ResourceSqlDatabaseInstance(),
 			"google_sql_ssl_cert":                           ResourceSqlSslCert(),
 			"google_sql_user":                               ResourceSqlUser(),
@@ -1709,7 +1714,7 @@ func ResourceMapWithErrors() (map[string]*schema.Resource, error) {
 			"google_runtimeconfig_variable":                 runtimeconfig.ResourceRuntimeconfigVariable(),
 			"google_service_account":                        resourcemanager.ResourceGoogleServiceAccount(),
 			"google_service_account_key":                    resourcemanager.ResourceGoogleServiceAccountKey(),
-			"google_service_networking_peered_dns_domain":   ResourceGoogleServiceNetworkingPeeredDNSDomain(),
+			"google_service_networking_peered_dns_domain":   servicenetworking.ResourceGoogleServiceNetworkingPeeredDNSDomain(),
 			"google_storage_bucket":                         ResourceStorageBucket(),
 			"google_storage_bucket_acl":                     ResourceStorageBucketAcl(),
 			"google_storage_bucket_object":                  ResourceStorageBucketObject(),
