+++ google-beta/resource_composer_environment_test.go
@@ -6,6 +6,7 @@ import (
 	"context"
 	"fmt"
 	"github.com/hashicorp/terraform-provider-google-beta/google-beta/acctest"
+	"github.com/hashicorp/terraform-provider-google-beta/google-beta/services/composer"
 	tpgcompute "github.com/hashicorp/terraform-provider-google-beta/google-beta/services/compute"
 	transport_tpg "github.com/hashicorp/terraform-provider-google-beta/google-beta/transport"
 	"testing"
@@ -38,36 +39,6 @@ func allComposerServiceAgents() []string {
 		"container-engine-robot",
 		"gcp-sa-artifactregistry",
 		"gcp-sa-pubsub",
-	}
-}
-
-func TestComposerImageVersionDiffSuppress(t *testing.T) {
-	t.Parallel()
-
-	cases := []struct {
-		name     string
-		old      string
-		new      string
-		expected bool
-	}{
-		{"matches", "composer-1.4.0-airflow-1.10.0", "composer-1.4.0-airflow-1.10.0", true},
-		{"preview matches", "composer-1.17.0-preview.0-airflow-2.0.1", "composer-1.17.0-preview.0-airflow-2.0.1", true},
-		{"old latest", "composer-latest-airflow-1.10.0", "composer-1.4.1-airflow-1.10.0", true},
-		{"new latest", "composer-1.4.1-airflow-1.10.0", "composer-latest-airflow-1.10.0", true},
-		{"composer major alias equivalent", "composer-1.4.0-airflow-1.10.0", "composer-1-airflow-1.10", true},
-		{"composer major alias different", "composer-1.4.0-airflow-2.1.4", "composer-2-airflow-2.2", false},
-		{"composer different", "composer-1.4.0-airflow-1.10.0", "composer-1.4.1-airflow-1.10.0", false},
-		{"airflow major alias equivalent", "composer-1.4.0-airflow-1.10.0", "composer-1.4.0-airflow-1", true},
-		{"airflow major alias different", "composer-1.4.0-airflow-1.10.0", "composer-1.4.0-airflow-2", false},
-		{"airflow major.minor alias equivalent", "composer-1.4.0-airflow-1.10.0", "composer-1.4.0-airflow-1.10", true},
-		{"airflow major.minor alias different", "composer-1.4.0-airflow-2.1.4", "composer-1.4.0-airflow-2.2", false},
-		{"airflow different", "composer-1.4.0-airflow-1.10.0", "composer-1.4.0-airflow-1.9.0", false},
-	}
-
-	for _, tc := range cases {
-		if actual := composerImageVersionDiffSuppress("", tc.old, tc.new, nil); actual != tc.expected {
-			t.Errorf("'%s' failed, expected %v but got %v", tc.name, tc.expected, actual)
-		}
 	}
 }
 
@@ -1032,15 +1003,15 @@ func testAccComposerEnvironmentDestroyProducer(t *testing.T) func(s *terraform.S
 			if len(idTokens) != 6 {
 				return fmt.Errorf("Invalid ID %q, expected format projects/{project}/regions/{region}/environments/{environment}", rs.Primary.ID)
 			}
-			envName := &composerEnvironmentName{
+			envName := &composer.ComposerEnvironmentName{
 				Project:     idTokens[1],
 				Region:      idTokens[3],
 				Environment: idTokens[5],
 			}
 
-			_, err := config.NewComposerClient(config.UserAgent).Projects.Locations.Environments.Get(envName.resourceName()).Do()
+			_, err := config.NewComposerClient(config.UserAgent).Projects.Locations.Environments.Get(envName.ResourceName()).Do()
 			if err == nil {
-				return fmt.Errorf("environment %s still exists", envName.resourceName())
+				return fmt.Errorf("environment %s still exists", envName.ResourceName())
 			}
 		}
