+++ google-beta/services/cloudrun/data_source_cloud_run_locations.go
@@ -1,6 +1,6 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
-package google
+package cloudrun
 
 import (
 	"fmt"
