+++ google-beta/services/composer/data_source_google_composer_environment.go
@@ -1,6 +1,6 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
-package google
+package composer
 
 import (
 	"fmt"
