+++ google-beta/services/container/resource_container_node_pool_migrate.go
@@ -1,6 +1,6 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
-package google
+package container
 
 import (
 	"fmt"
