+++ google-beta/services/container/state_util.go
@@ -1,6 +1,6 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
-package google
+package container
 
 // A StateType represents the specific type of resting state that a state value
 // is.
