+++ google-beta/services/containeranalysis/data_source_container_registry_image.go
@@ -1,6 +1,6 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
-package google
+package containeranalysis
 
 import (
 	"fmt"
