+++ google-beta/services/containerattached/data_source_google_container_attached_install_manifest.go
@@ -1,6 +1,6 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
-package google
+package containerattached
 
 import (
 	"fmt"
@@ -11,9 +11,9 @@ import (
 	transport_tpg "github.com/hashicorp/terraform-provider-google-beta/google-beta/transport"
 )
 
-func DataSourceGoogleContainerAttachedVersions() *schema.Resource {
+func DataSourceGoogleContainerAttachedInstallManifest() *schema.Resource {
 	return &schema.Resource{
-		Read: dataSourceGoogleContainerAttachedVersionsRead,
+		Read: dataSourceGoogleContainerAttachedInstallManifestRead,
 		Schema: map[string]*schema.Schema{
 			"project": {
 				Type:     schema.TypeString,
@@ -23,21 +23,31 @@ func DataSourceGoogleContainerAttachedVersions() *schema.Resource {
 				Type:     schema.TypeString,
 				Required: true,
 			},
-			"valid_versions": {
-				Type:     schema.TypeList,
+			"cluster_id": {
+				Type:     schema.TypeString,
+				Required: true,
+			},
+			"platform_version": {
+				Type:     schema.TypeString,
+				Required: true,
+			},
+			"manifest": {
+				Type:     schema.TypeString,
 				Computed: true,
-				Elem:     &schema.Schema{Type: schema.TypeString},
 			},
 		},
 	}
 }
 
-func dataSourceGoogleContainerAttachedVersionsRead(d *schema.ResourceData, meta interface{}) error {
+func dataSourceGoogleContainerAttachedInstallManifestRead(d *schema.ResourceData, meta interface{}) error {
 	config := meta.(*transport_tpg.Config)
 	userAgent, err := tpgresource.GenerateUserAgentString(d, config.UserAgent)
 	if err != nil {
 		return err
 	}
+
+	clusterId := d.Get("cluster_id").(string)
+	platformVersion := d.Get("platform_version").(string)
 
 	project, err := tpgresource.GetProject(d, config)
 	if err != nil {
@@ -52,7 +62,15 @@ func dataSourceGoogleContainerAttachedVersionsRead(d *schema.ResourceData, meta 
 		return fmt.Errorf("Cannot determine location: set location in this data source or at provider-level")
 	}
 
-	url, err := tpgresource.ReplaceVars(d, config, "{{ContainerAttachedBasePath}}projects/{{project}}/locations/{{location}}/attachedServerConfig")
+	url, err := tpgresource.ReplaceVars(d, config, "{{ContainerAttachedBasePath}}projects/{{project}}/locations/{{location}}:generateAttachedClusterInstallManifest")
+	if err != nil {
+		return err
+	}
+	params := map[string]string{
+		"attached_cluster_id": clusterId,
+		"platform_version":    platformVersion,
+	}
+	url, err = transport_tpg.AddQueryParams(url, params)
 	if err != nil {
 		return err
 	}
@@ -66,13 +84,9 @@ func dataSourceGoogleContainerAttachedVersionsRead(d *schema.ResourceData, meta 
 	if err != nil {
 		return err
 	}
-	var validVersions []string
-	for _, v := range res["validVersions"].([]interface{}) {
-		vm := v.(map[string]interface{})
-		validVersions = append(validVersions, vm["version"].(string))
-	}
-	if err := d.Set("valid_versions", validVersions); err != nil {
-		return err
+
+	if err := d.Set("manifest", res["manifest"]); err != nil {
+		return fmt.Errorf("Error setting manifest: %s", err)
 	}
 
 	d.SetId(time.Now().UTC().String())
