+++ google-beta/services/containeraws/data_source_google_container_aws_versions.go
@@ -1,6 +1,6 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
-package google
+package containeraws
 
 import (
 	"fmt"
@@ -11,43 +11,38 @@ import (
 	transport_tpg "github.com/hashicorp/terraform-provider-google-beta/google-beta/transport"
 )
 
-func DataSourceGoogleContainerAttachedInstallManifest() *schema.Resource {
+func DataSourceGoogleContainerAwsVersions() *schema.Resource {
 	return &schema.Resource{
-		Read: dataSourceGoogleContainerAttachedInstallManifestRead,
+		Read: dataSourceGoogleContainerAwsVersionsRead,
 		Schema: map[string]*schema.Schema{
 			"project": {
 				Type:     schema.TypeString,
-				Required: true,
+				Optional: true,
 			},
 			"location": {
 				Type:     schema.TypeString,
-				Required: true,
+				Optional: true,
 			},
-			"cluster_id": {
-				Type:     schema.TypeString,
-				Required: true,
-			},
-			"platform_version": {
-				Type:     schema.TypeString,
-				Required: true,
-			},
-			"manifest": {
-				Type:     schema.TypeString,
+			"valid_versions": {
+				Type:     schema.TypeList,
 				Computed: true,
+				Elem:     &schema.Schema{Type: schema.TypeString},
+			},
+			"supported_regions": {
+				Type:     schema.TypeList,
+				Computed: true,
+				Elem:     &schema.Schema{Type: schema.TypeString},
 			},
 		},
 	}
 }
 
-func dataSourceGoogleContainerAttachedInstallManifestRead(d *schema.ResourceData, meta interface{}) error {
+func dataSourceGoogleContainerAwsVersionsRead(d *schema.ResourceData, meta interface{}) error {
 	config := meta.(*transport_tpg.Config)
 	userAgent, err := tpgresource.GenerateUserAgentString(d, config.UserAgent)
 	if err != nil {
 		return err
 	}
-
-	clusterId := d.Get("cluster_id").(string)
-	platformVersion := d.Get("platform_version").(string)
 
 	project, err := tpgresource.GetProject(d, config)
 	if err != nil {
@@ -62,15 +57,7 @@ func dataSourceGoogleContainerAttachedInstallManifestRead(d *schema.ResourceData
 		return fmt.Errorf("Cannot determine location: set location in this data source or at provider-level")
 	}
 
-	url, err := tpgresource.ReplaceVars(d, config, "{{ContainerAttachedBasePath}}projects/{{project}}/locations/{{location}}:generateAttachedClusterInstallManifest")
-	if err != nil {
-		return err
-	}
-	params := map[string]string{
-		"attached_cluster_id": clusterId,
-		"platform_version":    platformVersion,
-	}
-	url, err = transport_tpg.AddQueryParams(url, params)
+	url, err := tpgresource.ReplaceVars(d, config, "{{ContainerAwsBasePath}}projects/{{project}}/locations/{{location}}/awsServerConfig")
 	if err != nil {
 		return err
 	}
@@ -84,9 +71,16 @@ func dataSourceGoogleContainerAttachedInstallManifestRead(d *schema.ResourceData
 	if err != nil {
 		return err
 	}
-
-	if err := d.Set("manifest", res["manifest"]); err != nil {
-		return fmt.Errorf("Error setting manifest: %s", err)
+	if err := d.Set("supported_regions", res["supportedAwsRegions"]); err != nil {
+		return err
+	}
+	var validVersions []string
+	for _, v := range res["validVersions"].([]interface{}) {
+		vm := v.(map[string]interface{})
+		validVersions = append(validVersions, vm["version"].(string))
+	}
+	if err := d.Set("valid_versions", validVersions); err != nil {
+		return err
 	}
 
 	d.SetId(time.Now().UTC().String())
