+++ google-beta/services/containerazure/data_source_google_container_azure_versions.go
@@ -1,6 +1,6 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
-package google
+package containerazure
 
 import (
 	"fmt"
@@ -11,9 +11,9 @@ import (
 	transport_tpg "github.com/hashicorp/terraform-provider-google-beta/google-beta/transport"
 )
 
-func DataSourceGoogleContainerAwsVersions() *schema.Resource {
+func DataSourceGoogleContainerAzureVersions() *schema.Resource {
 	return &schema.Resource{
-		Read: dataSourceGoogleContainerAwsVersionsRead,
+		Read: dataSourceGoogleContainerAzureVersionsRead,
 		Schema: map[string]*schema.Schema{
 			"project": {
 				Type:     schema.TypeString,
@@ -37,7 +37,7 @@ func DataSourceGoogleContainerAwsVersions() *schema.Resource {
 	}
 }
 
-func dataSourceGoogleContainerAwsVersionsRead(d *schema.ResourceData, meta interface{}) error {
+func dataSourceGoogleContainerAzureVersionsRead(d *schema.ResourceData, meta interface{}) error {
 	config := meta.(*transport_tpg.Config)
 	userAgent, err := tpgresource.GenerateUserAgentString(d, config.UserAgent)
 	if err != nil {
@@ -57,7 +57,7 @@ func dataSourceGoogleContainerAwsVersionsRead(d *schema.ResourceData, meta inter
 		return fmt.Errorf("Cannot determine location: set location in this data source or at provider-level")
 	}
 
-	url, err := tpgresource.ReplaceVars(d, config, "{{ContainerAwsBasePath}}projects/{{project}}/locations/{{location}}/awsServerConfig")
+	url, err := tpgresource.ReplaceVars(d, config, "{{ContainerAzureBasePath}}projects/{{project}}/locations/{{location}}/azureServerConfig")
 	if err != nil {
 		return err
 	}
@@ -71,7 +71,7 @@ func dataSourceGoogleContainerAwsVersionsRead(d *schema.ResourceData, meta inter
 	if err != nil {
 		return err
 	}
-	if err := d.Set("supported_regions", res["supportedAwsRegions"]); err != nil {
+	if err := d.Set("supported_regions", res["supportedAzureRegions"]); err != nil {
 		return err
 	}
 	var validVersions []string
