+++ google-beta/services/servicenetworking/data_source_google_service_networking_peered_dns_domain.go
@@ -1,6 +1,6 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
-package google
+package servicenetworking
 
 import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
