+++ google-beta/services/servicenetworking/resource_google_service_networking_peered_dns_domain.go
@@ -1,6 +1,6 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
-package google
+package servicenetworking
 
 import (
 	"fmt"
